+++ validator/keymanager/remote-web3signer/BUILD.bazel
@@ -20,6 +20,7 @@ go_library(
         "//validator/keymanager:go_default_library",
         "//validator/keymanager/remote-web3signer/internal:go_default_library",
         "//validator/keymanager/remote-web3signer/v1:go_default_library",
+        "//validator/keymanager/remote-utils:go_default_library",
         "@com_github_ethereum_go_ethereum//common/hexutil:go_default_library",
         "@com_github_go_playground_validator_v10//:go_default_library",
         "@com_github_logrusorgru_aurora//:go_default_library",
