+++ validator/keymanager/types.go
@@ -5,13 +5,11 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/logrusorgru/aurora"
 	"github.com/prysmaticlabs/prysm/async/event"
 	fieldparams "github.com/prysmaticlabs/prysm/config/fieldparams"
 	"github.com/prysmaticlabs/prysm/crypto/bls"
 	ethpbservice "github.com/prysmaticlabs/prysm/proto/eth/service"
 	validatorpb "github.com/prysmaticlabs/prysm/proto/prysm/v1alpha1/validator-client"
-	"github.com/prysmaticlabs/prysm/validator/accounts/petnames"
 )
 
 // IKeymanager defines a general keymanager interface for Prysm wallets.
@@ -134,19 +132,5 @@ func ParseKind(k string) (Kind, error) {
 		return Web3Signer, nil
 	default:
 		return 0, fmt.Errorf("%s is not an allowed keymanager", k)
-	}
-}
-
-// DisplayRemotePublicKeys prints remote public keys to stdout.
-func DisplayRemotePublicKeys(validatingPubKeys [][48]byte) {
-	au := aurora.NewAurora(true)
-	for i := 0; i < len(validatingPubKeys); i++ {
-		fmt.Println("")
-		fmt.Printf(
-			"%s\n", au.BrightGreen(petnames.DeterministicName(validatingPubKeys[i][:], "-")).Bold(),
-		)
-		// Retrieve the validating key account metadata.
-		fmt.Printf("%s %#x\n", au.BrightCyan("[validating public key]").Bold(), validatingPubKeys[i])
-		fmt.Println(" ")
 	}
 }
